Add benchmark and refactor
